+++ src/main/java/net/gegy1000/terrarium/server/map/cover/generator/FloodedGrasslandCover.java
@@ -105,7 +105,7 @@ public class FloodedGrasslandCover extends CoverGenerator {
     }
 
     @Override
-    protected IBlockState getFillerAt(Random random, int x, int z, byte slope) {
+    protected IBlockState getFillerAt(Random random, int x, int z, int slope) {
         return COARSE_DIRT;
     }
 }
