+++ src/main/java/net/gegy1000/terrarium/server/map/cover/generator/WaterCover.java
@@ -31,7 +31,7 @@ public class WaterCover extends CoverGenerator {
     }
 
     @Override
-    protected IBlockState getCoverAt(Random random, int x, int z, byte slope) {
+    protected IBlockState getCoverAt(Random random, int x, int z, int slope) {
         return WATER;
     }
